upgraded the Math attribute and math expression so that they can handle doubles; it will increase the precision as required by BG (mantis 1840)
git-svn-id: https://svn.opendtect.org/od/trunk@31607 a985855d-d384-4b45-ae33-31a410998c3c

+++ include/Attributes/mathattrib.h
@@ -60,9 +60,9 @@ private:
     ObjectSet<const DataHolder>	inputdata_;
     TypeSet<int>		inputidxs_;
 
-    TypeSet<float>		csts_;
+    TypeSet<double>		csts_;
     MathExpression*		expression_;
-    TypeSet<float>		recstartvals_;
+    TypeSet<double>		recstartvals_;
     float			recstartpos_;
     Interval<float>		desintv_;
     Interval<int>		reqintv_;
